config pacman/paru & ly [+numlock on start]

+++ init.sh
@@ -9,9 +9,13 @@ sudo pacman -S --noconfirm --needed --quiet git zsh stow
 
 stow bob/ fonts/ i3/ nchat/ ncspot/ nvim/ octave/ pip/ rofi/ scripts/ sound_effects/ tmux/ wallpapers/ wezterm/ zsh/
 
+# Default wallpaper
+echo "$HOME/dotconfigs/wallpapers/.local/share/wallpapers/3DAbstract/nku2ak42wzg31.png" > $HOME/.background
+
 (cat /etc/passwd | grep "$USER" | awk -F':' '{print $7}' | grep "zsh" > /dev/null) || chsh -s "/usr/bin/zsh" "$USER"
 
-REQUIRED_PACMAN_PACKAGES="ardour bat bob brave calf curl dolphin dragonfly-bin eza fd feh firefox fzf gcc gdb guitarix i3 lazygit lua luarocks ly maim make mesa mesa-demos mesa-utils nano numlockx onlyoffice openssh opera pacman pavucontrol pipewire pipewire-alsa pipewire-autostart pipewire-pulse pipewire-zeroconf qpwgraph ripgrep rofi steam sublime-text-4 thefuck tmux vim wezterm wget wine xclip xorg zoxide zoxide"
+REQUIRED_PACMAN_PACKAGES="ardour bat bob calf curl dolphin dragonfly-reverb eza fd feh firefox fzf gcc gdb guitarix i3 lazygit lolcat lua luarocks ly maim make mesa mesa-demos mesa-utils nano numlockx onlyoffice openssh pavucontrol pipewire pipewire-alsa pipewire-autostart pipewire-pulse pipewire-zeroconf qpwgraph ripgrep rofi screenfetch steam thefuck tmux vim wezterm wget wine xclip xorg zoxide"
+REQUIRED_AUR_PACKAGES="brave systemd-numlockontty opera sublime-text-4"
 
 function existCommand { command -v "$1" > /dev/null; }
 
@@ -19,17 +23,27 @@ existCommand "paru"		|| (sudo pacman -S --needed base-devel && git clone https:/
 existCommand "paruz"	|| (git clone https://github.com/joehillen/paruz.git $HOME/.local/builds/paruz && cd $HOME/.local/builds/paruz && sudo make install && cd - > /dev/null)
 existCommand "nvm"		|| (PROFILE=/dev/null bash -c 'curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.39.7/install.sh | bash')
 
-eval "paru -Syy --noconfirm --needed --quiet $REQUIRED_PACMAN_PACKAGES"
+# Config pacman & paru
+sudo sed -i.bak "s/#BottomUp/BottomUp/" /etc/paru.conf
+sudo sed -i.bak "s/#Color/Color/" /etc/pacman.conf
+sudo sed -i.bak -z "s/#\[multilib\]\n*\r*#Include/[multilib]\nInclude/" /etc/pacman.conf
 
 # Create dirs
 mkdir -p $HOME/.local/{logs,share/bob}
 
-sudo systemctl enable ly
-[ ! -f "$HOME/.ssh/id_ed25519" ] && ssh-keygen -t ed25519 -q -f "$HOME/.ssh/id_ed25519" -N ""
-
 zsh -ic "
+  eval \"paru -Syy --noconfirm --needed --quiet $REQUIRED_PACMAN_PACKAGES\"
+  eval \"paru -Syy --noconfirm --needed --quiet $REQUIRED_AUR_PACKAGES\"
+  
+  # Config ly
+  sudo sed -i.bak -z \"s/#animate = false/animate = true/\" /etc/ly/config.ini
+  sudo sed -i.bak -z \"s/#bigclock/bigclock/\" /etc/ly/config.ini
+  sudo systemctl enable ly numLockOnTty.service
+
+  [ ! -f \"$HOME/.ssh/id_ed25519\" ] && ssh-keygen -t ed25519 -q -f \"$HOME/.ssh/id_ed25519\" -N \"\"
+
 	existCommand \"nvm\" && nvm install $NODE_VERSION
 	existCommand \"bob\" && bob install latest && bob use latest || printf \"${RED}Command bob not found...\"
 	existCommand \"npm\" && npm i -g neovim nodemon pnpm spottydl ts-node typescript yarn || printf \"${RED}Command npm not found...\"
-  reboot
+  [ -z $1 ] && reboot
 "
